encoders/wav2vec2: undo the Wav2Vec2Config merge — keep config_wav2vec2.py
The previous commit (53f966ac99f) merged Wav2Vec2Config into model_wav2vec2.py
and used ``TYPE_CHECKING`` in the sibling files to dodge a circular import
that the merge itself created. The TYPE_CHECKING workaround was only needed
because the merge introduced the cycle — not for any structural reason.

Reverting the merge restores config_wav2vec2.py as a real leaf module
(no internal imports), and the siblings import Wav2Vec2Config from it
the normal way — no TYPE_CHECKING, no circular. ``__init__.py`` stays
deleted (its removal was the orthogonal pattern-alignment goal).

External importers (3 sites) now use two-line imports, e.g.::

    from ...encoders.wav2vec2.config_wav2vec2 import Wav2Vec2Config
    from ...encoders.wav2vec2.model_wav2vec2 import Wav2Vec2Encoder

This doesn't fully mirror clip/t5/umt5 (which keep Config + Encoder in one
``model_*.py``) — wav2vec2's split-into-5-files structure is honest about
the larger module surface. Matching clip's single-file layout would require
folding ``encoder_wav2vec2.py`` and ``feature_extractor_wav2vec2.py`` into
``model_wav2vec2.py`` too (file balloons to ~750 lines), not just the Config.
